+++ core-application/src/main/java/gohigher/user/port/in/token/TokenCommandPort.java
@@ -1,4 +1,4 @@
-package gohigher.port.in.token;
+package gohigher.user.port.in.token;
 
 import java.util.Date;
