+++ in-adapter-api/src/main/java/gohigher/auth/support/LoginArgumentResolver.java
@@ -7,7 +7,7 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
-import gohigher.port.in.token.TokenCommandPort;
+import gohigher.user.port.in.token.TokenCommandPort;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
