+++ security/src/main/java/gohigher/jwt/JwtAuthFilter.java
@@ -14,10 +14,10 @@ import org.springframework.web.filter.OncePerRequestFilter;
 import gohigher.global.exception.GoHigherException;
 import gohigher.jwt.support.AuthorizationExtractor;
 import gohigher.jwt.support.RoleGrantedAuthority;
-import gohigher.port.in.UserQueryPort;
-import gohigher.port.in.token.TokenCommandPort;
 import gohigher.user.AuthErrorCode;
 import gohigher.user.User;
+import gohigher.user.port.in.UserQueryPort;
+import gohigher.user.port.in.token.TokenCommandPort;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
