+++ security/src/main/java/gohigher/oauth2/OauthUserService.java
@@ -12,9 +12,9 @@ import org.springframework.stereotype.Service;
 import gohigher.jwt.support.RoleGrantedAuthority;
 import gohigher.oauth2.user.OAuth2UserInfo;
 import gohigher.oauth2.user.OAuth2UserInfoFactory;
-import gohigher.usecase.UserCommandService;
 import gohigher.user.User;
 import gohigher.user.auth.Provider;
+import gohigher.user.usecase.UserCommandService;
 import lombok.RequiredArgsConstructor;
 
 @Service
